+++ src/components/styles/create_account.css
@@ -1,0 +1,69 @@
+
+.form[data-tab="register"]{
+    display: flex;
+    flex-direction: column;
+    gap: 1.2rem;
+    margin-top: 1.5rem;
+}
+
+.form[data-tab="register"] .input-field{
+    margin-bottom: 1.5rem;
+}
+
+.form[data-tab="register"] label{
+   display: block;
+   margin-bottom: 0.5rem;
+   color: #555;
+   font-size: 1rem;
+   font-weight: 500;
+}
+
+.form[data-tab="register"]input[type="text"], 
+.form[data-tab="register"]input[type="email"],
+.form[data-tab="register"]input[type="password"],
+.form[data-tab="register"]input[type="tel"],
+.form[data-tab="register"]select{
+   width: 100%;
+   padding: 5rem;
+   font-size: larger;
+   border: 1px solid #ddd;
+   border-radius: 8px;
+   box-shadow: inset 0 1px 3px rgba(0, 0, 0, 0.05);
+   transition: border-color 0.2s ease-in-out;
+
+}
+
+.form[data-tab="register"] input:focus,
+.form[data-tab="register"] select:focus{
+    border-color: #3498db;
+    outline: none;
+    box-shadow: inset 0 1px 3px rgba(0, 0, 0), 0 0 5px rgba(52, 152, 219, 0.3);
+}
+
+.form[data-tab="register"] select {
+    
+   box-sizing: border-box;
+   border-radius: 10px;
+    cursor: pointer;
+    width: 50%;
+    text-align: center;
+    padding: 0.3rem;
+}
+
+.form[data-tab="register"] select::-ms-expand {
+    display: none;
+}
+
+.form[data-tab="register"] .login-button {
+    padding: 1.1rem;
+    font-size: 1.1rem;
+    margin-top: 1.5rem;
+    border-radius: 8px;
+}
+
+.form[data-tab="register"] .error-message {
+    color: #e74c3c;
+    margin-bottom: 1rem;
+    font-size: 0.9rem;
+}
+
