+++ src/pages/styles/admin_panel.css
@@ -1,0 +1,29 @@
+.dashboard-container{
+display: flex;
+
+
+}
+
+.chart-wraper{
+    margin-left: 20px;
+    max-width: 500px;
+    padding: 20px;
+    color: #333;
+    background: white;
+    font-size: 24px;
+    border-radius: 60px;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+ }
+
+ .chart-wraper h2{
+   text-align: center;
+   padding-bottom: 20px;
+   
+ }
+
+ 
+
+ .chart-area{
+    width: 100%;
+    height: 300px;
+ }
